requirements.txt
streamlit
SpeechRecognition
openai-whisper
torch
transformers 